+++ app/modules/AdminModule/Presenters/Processes/ProcessesPresenter.php
@@ -2,6 +2,7 @@
 
 namespace App\Modules\AdminModule;
 
+use App\Components\Widgets\ProcessStatsWidget\ProcessStatsWidget;
 use App\Constants\Container\ProcessStatus;
 use App\Constants\Container\SystemGroups;
 use App\Core\DB\DatabaseRow;
@@ -25,6 +26,12 @@ class ProcessesPresenter extends AAdminPresenter {
     }
 
     public function renderDashboard() {}
+
+    protected function createComponentProcessStatsWidget() {
+        $widget = $this->componentFactory->createComponentInstanceByClassName(ProcessStatsWidget::class, [$this->processManager, $this->processInstanceManager]);
+
+        return $widget;
+    }
 
     public function renderList() {
         $links = [];